Add fdo build tag
