+++ pkg/clients/fdo/client.go
@@ -1,3 +1,6 @@
+//go:build fdo
+// +build fdo
+
 package fdo
 
 import (
