+++ pkg/clients/fdo/client_test.go
@@ -1,3 +1,6 @@
+//go:build fdo
+// +build fdo
+
 package fdo_test
 
 import (
