+++ pkg/routes/ownershipvoucher_test.go
@@ -1,3 +1,6 @@
+//go:build fdo
+// +build fdo
+
 package routes_test
 
 import (
