+++ pkg/services/ownershipvoucher_unit_test.go
@@ -1,3 +1,6 @@
+//go:build fdo
+// +build fdo
+
 package services
 
 import (
